Development

+++ app/src/main/java/org/eyeseetea/malariacare/LoginActivity.java
@@ -19,24 +19,37 @@
 
 package org.eyeseetea.malariacare;
 
+import android.app.AlertDialog;
 import android.app.LoaderManager.LoaderCallbacks;
+import android.content.Context;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.Loader;
 import android.content.SharedPreferences;
 import android.database.Cursor;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
+import android.text.Html;
+import android.text.SpannableString;
+import android.text.method.LinkMovementMethod;
+import android.text.util.Linkify;
+import android.util.Log;
 import android.widget.EditText;
+import android.widget.TextView;
 
 import com.squareup.otto.Subscribe;
 
 import org.eyeseetea.malariacare.database.iomodules.dhis.importer.PullController;
 import org.eyeseetea.malariacare.database.model.User;
 import org.eyeseetea.malariacare.database.utils.PopulateDB;
+import org.eyeseetea.malariacare.database.utils.PreferencesState;
 import org.eyeseetea.malariacare.database.utils.Session;
+import org.eyeseetea.malariacare.utils.AUtils;
 import org.hisp.dhis.android.sdk.job.NetworkJob;
 import org.hisp.dhis.android.sdk.persistence.Dhis2Application;
 import org.hisp.dhis.android.sdk.persistence.preferences.ResourceType;
+
+import java.io.InputStream;
 
 /**
  * Login Screen.
@@ -44,6 +57,7 @@ import org.hisp.dhis.android.sdk.persistence.preferences.ResourceType;
  */
 public class LoginActivity extends org.hisp.dhis.android.sdk.ui.activities.LoginActivity implements LoaderCallbacks<Cursor> {
 
+    private static final String TAG="LoginActivity";
     /**
      * DHIS server URL
      */
@@ -61,6 +75,7 @@ public class LoginActivity extends org.hisp.dhis.android.sdk.ui.activities.Login
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        Log.d(TAG, "onCreate");
         super.onCreate(savedInstanceState);
         SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
         if (User.getLoggedUser() != null && !ProgressActivity.PULL_CANCEL &&  sharedPreferences.getBoolean(getApplicationContext().getResources().getString(R.string.pull_metadata),false)) {
@@ -88,6 +103,61 @@ public class LoginActivity extends org.hisp.dhis.android.sdk.ui.activities.Login
 
     }
 
+    /**
+     * Shows an alert dialog asking for acceptance of the EULA terms. If ok calls login function, do nothing otherwise
+     * @param titleId
+     * @param rawId
+     * @param context
+     */
+    public void askEula(int titleId, int rawId, final Context context){
+        InputStream message = context.getResources().openRawResource(rawId);
+        String stringMessage = AUtils.convertFromInputStreamToString(message).toString();
+        final SpannableString linkedMessage = new SpannableString(Html.fromHtml(stringMessage));
+        Linkify.addLinks(linkedMessage, Linkify.EMAIL_ADDRESSES | Linkify.WEB_URLS);
+
+        AlertDialog dialog = new AlertDialog.Builder(context)
+                .setTitle(context.getString(titleId))
+                .setMessage(linkedMessage)
+                .setNeutralButton(android.R.string.ok, new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        rememberEulaAccepted(context);
+                        loginToDhis(serverUrl,username,password);
+                    }
+                })
+                .setNegativeButton(android.R.string.no, null).create();
+        dialog.show();
+        ((TextView)dialog.findViewById(android.R.id.message)).setMovementMethod(LinkMovementMethod.getInstance());
+    }
+
+    /**
+     * Save a preference to remember that EULA was already accepted
+     * @param context
+     */
+    public void rememberEulaAccepted(Context context){
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context);
+        SharedPreferences.Editor editor = sharedPreferences.edit();
+        editor.putBoolean(getString(R.string.eula_accepted), true);
+        editor.commit();
+    }
+
+    /**
+     * User SDK function to login
+     * @param serverUrl
+     * @param username
+     * @param password
+     */
+    public void loginToDhis(String serverUrl, String username, String password){
+        //Delegate real login attempt to parent in sdk
+        super.login(serverUrl, username, password);
+    }
+
+    /**
+     * Ask for EULA acceptance if this is the first time user login to the server, otherwise login
+     * @param serverUrl
+     * @param username
+     * @param password
+     */
     @Override
     public void login(String serverUrl, String username, String password) {
         //This method is overriden to capture credentials data
@@ -95,8 +165,13 @@ public class LoginActivity extends org.hisp.dhis.android.sdk.ui.activities.Login
         this.username=username;
         this.password=password;
 
-        //Delegate real login attempt to parent
-        super.login(serverUrl,username,password);
+
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        if (!sharedPreferences.getBoolean(getString(R.string.eula_accepted), false)) {
+            askEula(R.string.settings_menu_eula, R.raw.eula, LoginActivity.this);
+        } else {
+            loginToDhis(serverUrl, username, password);
+        }
     }
 
     @Subscribe
@@ -104,8 +179,9 @@ public class LoginActivity extends org.hisp.dhis.android.sdk.ui.activities.Login
         if(result!=null && result.getResourceType().equals(ResourceType.USERS)) {
             if(result.getResponseHolder().getApiException() == null) {
                 saveUserDetails();
-                //FIXME remove when create survey is fixed
-//                populateFromAssets();
+
+                populateFromAssetsIfRequired();
+
                 launchMainActivity();
             } else {
                 onLoginFail(result.getResponseHolder().getApiException());
@@ -116,15 +192,20 @@ public class LoginActivity extends org.hisp.dhis.android.sdk.ui.activities.Login
     /**
      * Utility method to use while developing to avoid a real pull
      */
-    private void populateFromAssets() {
+    private void populateFromAssetsIfRequired() {
+        //From server -> done
+        if(PreferencesState.getInstance().getPullFromServer()) {
+            return;
+        }
+
+        //Populate locally
         try{
+            PullController.getInstance().wipeDatabase();
             User user = new User();
             user.save();
             Session.setUser(user);
-            PullController.getInstance().wipeDatabase();
             PopulateDB.populateDB(getAssets());
         }catch(Exception ex){
-
         }
     }
 
@@ -138,6 +219,18 @@ public class LoginActivity extends org.hisp.dhis.android.sdk.ui.activities.Login
         editor.putString(getString(R.string.dhis_user), this.username);
         editor.putString(getString(R.string.dhis_password), this.password);
         editor.commit();
+    }
+
+    /**
+     * LoginActivity does NOT admin going backwads since it is always the first activity.
+     * Thus onBackPressed closes the app
+     */
+    @Override
+    public void onBackPressed(){
+        Intent intent = new Intent(Intent.ACTION_MAIN);
+        intent.addCategory(Intent.CATEGORY_HOME);
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+        startActivity(intent);
     }
 
 }
